compiler/WeatherKitProvider/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pip==22.3
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
setuptools==65.5.0
urllib3==1.26.14
wheel==0.37.1
cryptography==39.0.1
